Docker/requirements.txt
#donut-shellcode
#psycopg
django
channels
requests
apscheduler
djangorestframework
redis
django-redis
rsa
daphne
channels_redis
python-telegram-bot
pycryptodome
cryptography
jinja2
unicorn
pyyaml
chardet
geoip2
dnslib
asgi_redis
tldextract
pluginbase
mitmproxy
lxml
dnspython
gevent
zope-interface
langchain-community
langchain-openai
langchain-core
langgraph
elasticsearch-dsl